ref: handle the amount of arguments per function

+++ server/include/utils.h
@@ -11,7 +11,7 @@ typedef struct {
 
 string_vec_t *string_vec_init(void);
 string_vec_t *string_split(char *str, char *del);
-size_t string_split_amount(char *str, char *del);
+int string_split_amount(char *str, char *del);
 void string_vec_free(string_vec_t *vec);
 
 int strcmp_end(const char *str, const char *end);
